requirements.txt
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.30
botocore==1.17.30
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
deap==1.3.1
docutils==0.15.2
emoji==0.5.4
Flask==1.1.2
Flask-Cors==3.0.9
gensim==3.8.3
gevent==20.6.2
google==2.0.3
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.30.0
grpcio-tools==1.30.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mysql-connector==2.2.9
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
Pillow==7.2.0
protobuf==3.12.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
smart-open==2.1.0
soupsieve==2.0.1
stopit==1.1.2
threadpoolctl==2.1.0
TPOT==0.11.5
tqdm==4.48.2
tweepy==3.8.0
update-checker==0.18.0
urllib3==1.25.9
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
